requirements.txt
flake8
black
fastapi
uvicorn
pydantic
requests
boto3
pymongo
motor
beanie
python-dotenv
python-multipart
selenium
webdriver-manager
pdfkit
ffmpeg-python
python-jose
bcrypt
redis
